[CCUBE-1662][RL] Extract conditionally rendered parent
